testy/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-results
celery-progress==0.1.3
    # via -r requirements.in
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-js-asset
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-celery-results==2.4.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-mptt
django-mptt==0.13.4
    # via -r requirements.in
django-simple-history==3.1.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
drf-yasg==1.21.4
    # via -r requirements.in
flake8==5.0.4
    # via flake8-copyright-validator
flake8-copyright-validator==0.0.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via flake8-copyright-validator
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
kombu==5.2.4
    # via celery
markupsafe==2.1.1
    # via jinja2
mccabe==0.7.0
    # via flake8
packaging==21.3
    # via drf-yasg
prompt-toolkit==3.0.36
    # via click-repl
psycopg2==2.9.3
    # via -r requirements.in
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via packaging
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   celery
    #   django
    #   djangorestframework
    #   drf-yasg
redis==4.4.0
    # via -r requirements.in
requests==2.28.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
six==1.16.0
    # via click-repl
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
